Lightgun offscreen adjustments
Core option that affects how the lightgun behaves when pointing offscreen:

Free: cursor can move freely along edges of screen when the lightgun is pointed offscreen. This is useful for games that don't require offscreen reloading or may be negatively affected by the the other settings (gunbustr). The GUN RELOAD input fires a shot at the screen's top-left corner.

Fixed (top left): when the cursor touches a min or max value, it is fixed to the top-left corner of the screen. This usually ensures proper reloading for games with offscreen reload. The GUN RELOAD button places the cursor in that same spot.

Fixed (bottom right): Same as top left, but this instead uses the bottom-right corner for the games that require reload but don't have working reload in the top-left corner with the lightgun input (lethalj, invasnab,...). The GUN RELOAD button fires a shot at the bottom-right corner with this selected.